Implement experimental re-exploration through fork
This solution is very ineffective as it calls fork recursively therefore
causing an ever increasing process tree to be created. Unfortunately, using
fork is necessary to reset SystemC. An alternative solution would be to
use IPC mechanisms to communicate new branches back to the parent
process. However, serializing klee:Exprs (as encapsulated in clover's
BitVector object) is difficult. In the meantime, the solution
implemented by this commit is good enough to test concolic execution
with simple program which just have two or three branches.