hifive-vp: Fix loading of ELF sections into Dram
Without this change, the ELF Loader is only capable of loading ELF
regions to the Flash memory segment. Attempting to process regions which
should be loaded into the DRAM segment [0x80000000, 0x80004000)
presently causes an exception to be raised. This prevents ELF binaries
with such regions from being executed.

For example, this prevents Zephyr from booting on the hifive-vp:

	$ readelf -a zephyr_hello_world.elf | grep 0x80000000
	  LOAD           0x000000 0x80000000 0x80000000 0x00000 0x00f60 RW  0x10
	$ hifive-vp zephyr_hello_world.elf
	[elf_loader] Section does not fit in target memory
	Error: (E549) uncaught exception: Tried loading invalid elf layout
	In file: ../../../src/sysc/kernel/sc_except.cpp:101

This commit adds support for ELF regions which should be loaded to the
DRAM Section by invoking loader.load_executable_image twice and silently
ignoring ELF regions which cannot be mapped to the given memory
segment. I would consider this a hotfix, the prober way to fix this
issue would probably be attaching the ELF Loader to the TLM Bus or
passing some sort of combined memory interface to the loader which
bundles the DRAM and Flash.

However, with this commit applied Zephyr boots again succesfully:

	$ hifive-vp zephyr_hello_world.elf
	*** Booting Zephyr OS build v2.7.99-644-g75cb6e649863  ***
	Hello World! hifive1