AbstractUART: Split code to stop threads from destructor
And instead stop the threads from the subclass destructor. This should
fix a race condition where poll would receive input from the input fd
after the subclass destructor has returned but before the AbstractUART
destructor has run. In this case, the receive thread would call
handle_input() after the subclass destructor has returned. In this case,
the call to handle_input() would be an invocation of a pure virtual
function which would result in an exception.