Remove all floating-point related code
Contrary to riscv-vp, we don't have any ambitions to support the
F-extensions in SymEx-VP at the moment and doing so it certainly
challenging due to state explosion issues induced by floating
point operations with symbolic values.

For this reason, this commit removes and disables all floating-point
related code. This allows us to remove the vendored softfloat library
and results in a significant build time reduction.