Remove syscall interception support
The syscall interception newer really worked with SymEx-VP since our
memory implementation is a HashMap and can thus not be passed to
sys.Init as is. For this reason, most intercepted syscalls caused
a null pointer dereference.

Furthermore, syscalls are nowadays no longer used to interact with
the symbolic execution engine as this is now achieved over MMIO using
the SymbolicCTRL peripheral. As such, the syscall interception should
not be needed anymore.

See e421dd1fd45a0106c778d09881d099906db28ddf

Fixes #13